- added energy deposit collection algorithms: 'all' and 'adjacent', messengers   and CexmcRunSObject were updated - better help formats in CexmcChargeExchangeReconstructorMessenger - all links to reference crystal selection algorithm were removed from   CexmcEventAction. Previous solution could change original edStore if inner   crystals had been chosen for referencing the crystal with maximum of absorbed   energy. Now processing of this algorithm is encapsulated in reconstructor   classes - TransformToAdjacentInnerCrystal() was moved from CexmcEnergyDepositDigitizer   to CexmcReconstructor - mkjob was enhanced to also make replay mode run jobs - a cast to size_t was needed in CexmcScenePrimitivesMessenger

+++ src/CexmcScenePrimitivesMessenger.cc
@@ -148,7 +148,7 @@ void  CexmcScenePrimitivesMessenger::SetNewValue( G4UIcommand *  cmd,
 
             iss >> name >> redOrString >> green >> blue >> opacity;
 
-            if ( std::isalpha( redOrString[ 0 ] ) )
+            if ( std::isalpha( redOrString[ size_t( 0 ) ] ) )
             {
                 G4Colour::GetColour( redOrString, colour );
             }
